requirements.txt
Django>=4.2,<5.0
Pillow>=9.0.0
python-decouple>=3.6
psycopg2-binary>=2.9.0
gunicorn>=20.0.0
whitenoise>=6.0.0
django-extensions>=3.2.0
django-ckeditor>=6.5.0
requests>=2.28.0
lxml>=4.6.0
django-robots>=5.0
django-crispy-forms>=1.14.0
crispy-bootstrap5>=0.7
dj-database-url>=1.0.0
